.h fix

+++ src/boot/ksdk1.1.0/devINA219.h
@@ -39,5 +39,5 @@
 void		initINA219(const uint8_t i2cAddress, uint16_t operatingVoltageMillivolts);
 WarpStatus	readSensorRegisterINA219(uint8_t deviceRegister, int numberOfBytes);
 WarpStatus	writeSensorRegisterINA219(uint8_t deviceRegister, uint16_t payloadBtye);
-WarpStatus	configureSensorINA219(uint16_t configBits, uint8_t calibrationValue);
+WarpStatus	configureSensorINA219(uint16_t configBits, uint16_t calibrationValue);
 void		printSensorDataINA219(bool hexModeFlag);
